+++ px_device_identity/log.py
@@ -1,17 +1,23 @@
-from os import environ
+import sys
+import syslog
 import logging
+from os import environ
+
 logging.basicConfig(level=environ.get("LOGLEVEL", "INFO"))
-log = logging.getLogger("LOG")
 
 class Logger:
-    def __init__(self, context):
+    def __init__(self, context, application="px-device-identity"):
         self.context = context
+        self.name = application
+        self.log = logging.getLogger(self.name)
 
     def info(self, message):
-        log.info(message)
+        self.log.info(message)
 
     def warning(self, message):
-        log.warn(message)
+        self.log.warn(message)
+        syslog.syslog(syslog.LOG_WARNING, message)
 
     def error(self, message):
-        log.error(message)
+        self.log.error(message)
+        syslog.syslog(syslog.LOG_ERR, message)